pyproject.toml
[tool.poetry]
name = "fastapi-unitofwork"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_unitofwork", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
pydantic-settings = "^2.1.0"
uvicorn = "^0.27.1"
aiosqlite = "^0.19.0"
alembic = "^1.13.1"
pyjwt = "^2.8.0"
pydantic = {extras = ["email"], version = "^2.6.1"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
